refactor(runtime,desktop): hoist history compact artifact helpers
Move loadHistoryCompactBlocksFromArtifacts and
persistHistoryCompactBlocksToArtifacts (with abort-cleanup tracking and
the loader source/kind filter) into @maka/runtime. Add an optional
abortSignal field to HistoryCompactWriteInput so the helper can abort a
write and delete every artifact it created. The desktop main file
becomes a thin re-export.
